4th week, started 5th

+++ gen_full_pdf.sh
@@ -4,10 +4,18 @@
 
 echo "Uniting screen pdfs..."
 cd "lection/$1/screen"
-pdfunite $(find . -type f -name '*.pdf' | sort) ../screen.pdf
-cd ..
-pdfunite conspect.pdf screen.pdf conspect_screen.pdf
-rm screen.pdf
+
+pdfs=$(find . -type f -name '*.pdf' | sort)
+if [[ -n "$pdfs" ]]; then
+  pdfunite "$pdfs" ../screen.pdf
+  cd ..
+  pdfunite conspect.pdf screen.pdf conspect_screen.pdf
+  rm screen.pdf
+else
+  cd ..
+  cp conspect.pdf conspect_screen.pdf
+fi
+
 cd ../..
 
 echo "Uniting all conspects pdfs..."
